requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
boto3==1.38.36
botocore==1.38.36
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
executing==2.2.0
google-auth==2.40.3
google-genai==1.20.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.86.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyzmq==27.0.0
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
